putting cylinder disks in place to work with unions and start debugging axis-aligned box calculation

+++ incl/libft.h
@@ -34,8 +34,8 @@ typedef struct s_list
 
 typedef struct s_range
 {
-	double	min;
-	double	max;
+	float	min;
+	float	max;
 }			t_range;
 
 int		ft_atoi(const char *str);
