+++ libft/srcs/gen_maths/ft_scale.c
@@ -6,15 +6,15 @@
 /*   By: djagusch <djagusch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/20 23:47:30 by djagusch          #+#    #+#             */
-/*   Updated: 2023/06/21 23:11:49 by djagusch         ###   ########.fr       */
+/*   Updated: 2023/07/08 12:45:57 by djagusch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_scale(double value, t_range range_old, t_range range_new)
+float	ft_scale(float value, t_range range_old, t_range range_new)
 {
-	double	scaled_val;
+	float	scaled_val;
 
 	scaled_val = (value - range_old.min);
 	scaled_val /= (range_old.max - range_old.min);
